refactor: change carriages structure
