+++ src/app/admin/admin.module.ts
@@ -2,8 +2,6 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { AdminRoutingModule } from './admin-routing.module';
-import { CarriagesComponent } from './pages/carriages/carriages.component';
-import { CarriageComponent } from './components/carriage/carriage.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
@@ -16,7 +14,7 @@ import { StationFormComponent } from './components/station-form/station-form.com
 import { StationMapComponent } from './components/station-map/station-map.component';
 
 @NgModule({
-  declarations: [CarriagesComponent, CarriageComponent, StationsComponent, StationListComponent, StationFormComponent, StationMapComponent],
+  declarations: [StationsComponent, StationListComponent, StationFormComponent, StationMapComponent],
   imports: [CommonModule, AdminRoutingModule, FormsModule, MatFormFieldModule, MatInputModule, MatButtonModule, MatExpansionModule, MatIconModule, MatSelectModule]
 })
 export class AdminModule {}
