--- src/app/admin/carriages/carriages.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  declarations: [],
-  imports: [CommonModule]
-})
-export class CarriagesModule {}
