+++ src/app/admin/carriages/components/carriage/carriage.component.scss
