+++ src/app/admin/carriages/components/carriage/carriage.component.ts
@@ -1,7 +1,10 @@
+import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 
 @Component({
   selector: 'app-carriage',
+  standalone: true,
+  imports: [CommonModule],
   styleUrl: './carriage.component.scss',
   templateUrl: './carriage.component.html'
 })
