--- src/app/admin/models/carriages.model.ts
@@ -1,11 +1,0 @@
-export interface Carriage {
-  code: string;
-  name: string;
-  rows: number;
-  leftSeats: number;
-  rightSeats: number;
-}
-
-export interface AuthResponse {
-  token: string;
-}
